requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.10.1
openai==1.3.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
python-multipart
google-generativeai

